implement global drawing method
