+++ figureSpacingProofUFO.py
@@ -21,17 +21,11 @@ import subprocess
 import drawBot as db
 
 from fontParts.fontshell import RFont
-from fontTools.pens.cocoaPen import CocoaPen
 
+from proofing_helpers.drawing import draw_glyph
 from proofing_helpers.files import get_ufo_paths
 from proofing_helpers.globals import FONT_MONO
 from proofing_helpers.stamps import timestamp
-
-
-def draw_glyph(glyph):
-    cpen = CocoaPen(glyph.getParent())
-    glyph.draw(cpen)
-    db.drawPath(cpen.path)
 
 
 def joinit(iterable, delimiter):
